requirements.txt
langchain>=0.0.300
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
tiktoken>=0.4.0
python-multipart>=0.0.6
fastapi>=0.95.2
uvicorn>=0.22.0
streamlit>=1.28.0
huggingface-hub
gunicorn
